requirements.txt
pandas
numpy
scikit-learn
plotly
streamlit
fastapi
uvicorn
joblib
prometheus-client
redis
kafka-python
python-dotenv
